[mlir][gpu] Fix error message on unknown CUDA error code.
This patch fixes the output of the error message that is printed when
the CUDA library cannot identity the error code. In that case, no error
message is provided by the library, and the previous implementation just
printed the content of a randomly initialized pointer. This patch
initializes the pointer to nullptr and only prints the content if that
has changed.

Reviewed By: Mogball

Differential Revision: https://reviews.llvm.org/D156791