[lldb][DWARF] Search for symbols in all external modules (#75927)
The way this code was updated in
dd9587795811ba21e6ca6ad52b4531e17e6babd6 meant that if the first module
did not have the symbol, the iteration stopped as returning true means
stop. So only if every module had the symbol would we find it, in the
last module.

Invert the condition to break when we find the first instance, which is
what the previous code did.