[llvm-exegesis] Add support for loading X86 segment registers (#76368)
This patch adds support for setting the X86 segment registers. These
registers are used in quite a few basic blocks in BHive and similar
datasets, so being able to set them is necessary to ensure consistent
runs as the live-in values of fs and gs can change across runs.

Fixes #76340.